+++ src/main/java/at/spengergasse/spring_thymeleaf/entities/TrinkflaschenRepository.java
@@ -2,6 +2,7 @@ package at.spengergasse.spring_thymeleaf.entities;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+// ✅ RICHTIG (passt zu Entity Long id)
 @Repository
-public interface TrinkflaschenRepository extends JpaRepository<Trinkflaschen, Integer>{
+public interface TrinkflaschenRepository extends JpaRepository<Trinkflaschen, Long> {
 }
